+++ core/src/test/scala/tech/beshu/ror/unit/acl/blocks/definitions/ldap/implementations/UnboundidLdapAuthorizationServiceInGroupsFromUserAttributeModeTests.scala
@@ -14,7 +14,7 @@
  *    You should have received a copy of the GNU General Public License
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
-package tech.beshu.ror.unit.acl.blocks.definitions
+package tech.beshu.ror.unit.acl.blocks.definitions.ldap.implementations
 
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.auto._
@@ -28,7 +28,7 @@ import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.Dn
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.LdapService.Name
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UnboundidLdapConnectionPoolProvider.LdapConnectionConfig
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UnboundidLdapConnectionPoolProvider.LdapConnectionConfig.{BindRequestUser, ConnectionMethod, LdapHost}
-import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UserGroupsSearchFilterConfig.UserGroupsSearchMode.{DefaultGroupSearch, GroupIdAttribute, GroupSearchFilter, NestedGroupsConfig, UniqueMemberAttribute}
+import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UserGroupsSearchFilterConfig.UserGroupsSearchMode._
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UserSearchFilterConfig.UserIdAttribute
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations._
 import tech.beshu.ror.accesscontrol.domain.{Group, PlainTextSecret, User}
@@ -40,25 +40,25 @@ import java.time.Clock
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
-class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeWhenUserIdAttributeIsUidTests
-  extends UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests {
+class UnboundidLdapAuthorizationServiceInGroupsFromUserAttributeModeWhenUserIdAttributeIsUidTests
+  extends UnboundidLdapAuthorizationServiceInGroupsFromUserAttributeModeTests {
 
   override protected val userIdAttribute: UserIdAttribute = UserIdAttribute.CustomAttribute("uid")
-  override protected val morganUserId: User.Id = User.Id("morgan")
+  override protected val jesusUserId: User.Id = User.Id("jesus")
   override protected val userSpeakerUserId: User.Id = User.Id("userSpeaker")
-  override protected val devitoUserId: User.Id = User.Id("devito")
+  override protected val spaghettiUserId: User.Id = User.Id("spaghetti")
 }
 
-class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeWhenUserIdAttributeIsCnTests
-  extends UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests {
+class UnboundidLdapAuthorizationServiceInGroupsFromUserAttributeModeWhenUserIdAttributeIsCnTests
+  extends UnboundidLdapAuthorizationServiceInGroupsFromUserAttributeModeTests {
 
   override protected val userIdAttribute: UserIdAttribute = UserIdAttribute.Cn
-  override protected val morganUserId: User.Id = User.Id("Morgan Freeman")
+  override protected val jesusUserId: User.Id = User.Id("Jesus Christ")
   override protected val userSpeakerUserId: User.Id = User.Id("UserSpeaker (ext)")
-  override protected val devitoUserId: User.Id = User.Id("Danny DeVito")
+  override protected val spaghettiUserId: User.Id = User.Id("Spaghetti Monster")
 }
 
-abstract class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests
+abstract class UnboundidLdapAuthorizationServiceInGroupsFromUserAttributeModeTests
   extends AnyWordSpec
     with BeforeAndAfterAll
     with Inside
@@ -79,9 +79,9 @@ abstract class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests
       "returns non empty set of groups" when {
         "user has groups" in {
           eventually {
-            peopleAndGroupsLdapAuthorizationService.groupsOf(morganUserId).runSyncUnsafe() should be {
-              UniqueList.of(group("groupAll"), group("group3"), group("group2"))
-            }
+            godAndRegionsLdapAuthorizationService.groupsOf(jesusUserId).runSyncUnsafe() should contain only(
+              group("europe"), group("north america"), group("south america"), group("africa")
+            )
           }
         }
       }
@@ -95,23 +95,23 @@ abstract class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests
       "returns empty set of groups" when {
         "user has no groups" in {
           eventually {
-            peopleAndGroupsLdapAuthorizationService.groupsOf(devitoUserId).runSyncUnsafe() should be(UniqueList.empty[Group])
+            godAndRegionsLdapAuthorizationService.groupsOf(spaghettiUserId).runSyncUnsafe() should be (UniqueList.empty[Group])
           }
         }
         "there is no user with given name" in {
           eventually {
-            peopleAndGroupsLdapAuthorizationService.groupsOf(User.Id("unknown")).runSyncUnsafe() should be(UniqueList.empty[Group])
+            godAndRegionsLdapAuthorizationService.groupsOf(User.Id("unknown")).runSyncUnsafe() should be(UniqueList.empty[Group])
           }
         }
       }
     }
   }
 
-  private def peopleAndGroupsLdapAuthorizationService = {
+  private def godAndRegionsLdapAuthorizationService = {
     implicit val clock: Clock = Clock.systemUTC()
     UnboundidLdapAuthorizationService
       .create(
-        Name("LDAP1"),
+        Name("ldap1"),
         ldapConnectionPoolProvider,
         LdapConnectionConfig(
           ConnectionMethod.SingleServer(
@@ -129,14 +129,13 @@ abstract class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests
           ),
           ignoreLdapConnectivityProblems = false
         ),
-        UserSearchFilterConfig(Dn("ou=People,dc=example,dc=com"), userIdAttribute),
+        UserSearchFilterConfig(Dn("ou=Gods,dc=example,dc=com"), userIdAttribute),
         UserGroupsSearchFilterConfig(
-          DefaultGroupSearch(
-            Dn("ou=Groups,dc=example,dc=com"),
-            GroupSearchFilter("(cn=*)"),
+          GroupsFromUserEntry(
+            Dn("ou=Regions,dc=example,dc=com"),
+            GroupSearchFilter("(objectClass=*)"),
             GroupIdAttribute("cn"),
-            UniqueMemberAttribute("uniqueMember"),
-            groupAttributeIsDN = true,
+            GroupsFromUserAttribute("title"),
           ),
           nestedGroupsConfig = None
         )
@@ -149,7 +148,7 @@ abstract class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests
     implicit val clock: Clock = Clock.systemUTC()
     UnboundidLdapAuthorizationService
       .create(
-        Name("LDAP1"),
+        Name("ldap1"),
         ldapConnectionPoolProvider,
         LdapConnectionConfig(
           ConnectionMethod.SingleServer(
@@ -169,17 +168,16 @@ abstract class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests
         ),
         UserSearchFilterConfig(Dn("ou=Users,dc=example,dc=com"), userIdAttribute),
         UserGroupsSearchFilterConfig(
-          DefaultGroupSearch(
+          GroupsFromUserEntry(
             Dn("ou=Roles,dc=example,dc=com"),
-            GroupSearchFilter("(cn=*)"),
+            GroupSearchFilter("(objectClass=*)"),
             GroupIdAttribute("cn"),
-            UniqueMemberAttribute("uniqueMember"),
-            groupAttributeIsDN = true,
+            GroupsFromUserAttribute("memberOf"),
           ),
           Some(NestedGroupsConfig(
             nestedLevels = 1,
             Dn("ou=Roles,dc=example,dc=com"),
-            GroupSearchFilter("(cn=*)"),
+            GroupSearchFilter("(objectClass=*)"),
             UniqueMemberAttribute("uniqueMember"),
             GroupIdAttribute("cn"),
           ))
@@ -190,7 +188,7 @@ abstract class UnboundidLdapAuthorizationServiceInDefaultGroupSearchModeTests
   }
 
   protected def userIdAttribute: UserIdAttribute
-  protected def morganUserId: User.Id
+  protected def jesusUserId: User.Id
   protected def userSpeakerUserId: User.Id
-  protected def devitoUserId: User.Id
+  protected def spaghettiUserId: User.Id
 }
