+++ core/src/test/scala/tech/beshu/ror/unit/acl/blocks/definitions/ldap/implementations/UnboundidLdapServiceNetworkRelatedTests.scala
@@ -14,26 +14,26 @@
  *    You should have received a copy of the GNU General Public License
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
-package tech.beshu.ror.unit.acl.blocks.definitions
+package tech.beshu.ror.unit.acl.blocks.definitions.ldap.implementations
 
+import cats.data._
 import com.dimafeng.testcontainers.{Container, ForAllTestContainer, MultipleContainers}
 import com.unboundid.ldap.sdk.LDAPSearchException
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.auto._
+import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
-import monix.execution.exceptions.ExecutionRejectedException
-import org.scalatest.concurrent.Eventually
+import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers._
-import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.{Assertion, BeforeAndAfterAll, BeforeAndAfterEach, Inside}
-import tech.beshu.ror.accesscontrol.blocks.definitions.CircuitBreakerConfig
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.LdapService.Name
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UnboundidLdapConnectionPoolProvider.LdapConnectionConfig
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UnboundidLdapConnectionPoolProvider.LdapConnectionConfig._
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UserSearchFilterConfig.UserIdAttribute.CustomAttribute
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations._
-import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.{CircuitBreakerLdapAuthenticationServiceDecorator, Dn, LdapAuthenticationService}
+import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.{Dn, LdapAuthenticationService}
 import tech.beshu.ror.accesscontrol.domain.{PlainTextSecret, User}
+import tech.beshu.ror.utils.ScalaOps.repeat
 import tech.beshu.ror.utils.SingletonLdapContainers
 import tech.beshu.ror.utils.containers.{LdapContainer, ToxiproxyContainer}
 
@@ -42,13 +42,12 @@ import scala.concurrent.duration._
 import scala.language.postfixOps
 import scala.reflect.ClassTag
 
-class LdapCircuitBreakerTests
-  extends AnyWordSpec
+class UnboundidLdapServiceNetworkRelatedTests
+  extends AnyFreeSpec
     with BeforeAndAfterAll
     with BeforeAndAfterEach
     with ForAllTestContainer
-    with Inside
-    with Eventually {
+    with Inside {
 
   private val ldap1ContainerWithToxiproxy = new ToxiproxyContainer(
     SingletonLdapContainers.ldap1,
@@ -71,51 +70,49 @@ class LdapCircuitBreakerTests
     ldap1ContainerWithToxiproxy.disableNetworkTimeout()
   }
 
-  "An CircuitBreaker decorated LdapAuthenticationService" should {
-    "close circuit breaker after 2 failed attempts" in {
-      val authenticationService = createCircuitBreakerDecoratedSimpleAuthenticationService()
-      eventually {
+  "An LdapAuthenticationService should" - {
+    "allow to authenticate user" - {
+      "after connection timeout and retry" in {
+        val authenticationService = createSimpleAuthenticationService()
         authenticationService.assertSuccessfulAuthentication
-        ldap1ContainerWithToxiproxy.disableNetwork()
-        authenticationService.assertFailedAuthentication[LDAPSearchException]
-        authenticationService.assertFailedAuthentication[LDAPSearchException]
-        authenticationService.assertFailedAuthentication[ExecutionRejectedException]
-        ldap1ContainerWithToxiproxy.enableNetwork()
-        authenticationService.assertFailedAuthentication[ExecutionRejectedException]
-      }
-    }
-    "close circuit breaker after 2 failed attempts, but open it later" in {
-      val authenticationService = createCircuitBreakerDecoratedSimpleAuthenticationService()
-      eventually {
-        authenticationService.assertSuccessfulAuthentication
-        ldap1ContainerWithToxiproxy.disableNetwork()
-        authenticationService.assertFailedAuthentication[LDAPSearchException]
-        ldap1ContainerWithToxiproxy.enableNetwork()
         ldap1ContainerWithToxiproxy.enableNetworkTimeout()
-        authenticationService.assertFailedAuthentication[LDAPSearchException]
-        authenticationService.assertFailedAuthentication[ExecutionRejectedException]
+        authenticationService.assertFailedAuthentication[LdapUnexpectedResult]
         ldap1ContainerWithToxiproxy.disableNetworkTimeout()
-        Thread.sleep(750)
-        authenticationService.assertSuccessfulAuthentication
         authenticationService.assertSuccessfulAuthentication
       }
-    }
-    "close circuit breaker after 2 failed attempts and keep it closed because of network issues" in {
-      val authenticationService = createCircuitBreakerDecoratedSimpleAuthenticationService()
-      eventually {
+      "after connection failure and retry" in {
+        val authenticationService = createSimpleAuthenticationService()
         authenticationService.assertSuccessfulAuthentication
         ldap1ContainerWithToxiproxy.disableNetwork()
         authenticationService.assertFailedAuthentication[LDAPSearchException]
-        authenticationService.assertFailedAuthentication[LDAPSearchException]
-        authenticationService.assertFailedAuthentication[ExecutionRejectedException]
-        Thread.sleep(750)
-        authenticationService.assertFailedAuthentication[LDAPSearchException]
-        authenticationService.assertFailedAuthentication[ExecutionRejectedException]
+        ldap1ContainerWithToxiproxy.enableNetwork()
+        authenticationService.assertSuccessfulAuthentication
+      }
+    }
+    "be able to work when" - {
+      "Round robin HA method is configured when" - {
+        "one of servers goes down" in {
+          def assertMorganCanAuthenticate(service: UnboundidLdapAuthenticationService) = {
+            service
+              .authenticate(User.Id("morgan"), PlainTextSecret("user1"))
+              .runSyncUnsafe() should be(true)
+          }
+          val service = createHaAuthenticationService()
+          (for {
+            _ <- repeat(maxRetries = 5, delay = 500 millis) {
+              Task(assertMorganCanAuthenticate(service))
+            }
+            _ <- Task(ldapContainerToStop.stop())
+            _ <- repeat(10, 500 millis) {
+              Task(assertMorganCanAuthenticate(service))
+            }
+          } yield ()).runSyncUnsafe()
+        }
       }
     }
   }
 
-  private implicit class LdapAuthenticationServiceOps(authenticationService: LdapAuthenticationService) {
+  implicit class LdapAuthenticationServiceOps(authenticationService: LdapAuthenticationService) {
     def assertSuccessfulAuthentication: Assertion = {
       authenticationService
         .authenticate(User.Id("morgan"), PlainTextSecret("user1"))
@@ -136,9 +133,7 @@ class LdapCircuitBreakerTests
         Name("my_ldap"),
         ldapConnectionPoolProvider,
         LdapConnectionConfig(
-          ConnectionMethod.SingleServer(
-            LdapHost.from(s"ldap://localhost:${ldap1ContainerWithToxiproxy.innerContainerMappedPort}").get
-          ),
+          ConnectionMethod.SingleServer(LdapHost.from(s"ldap://localhost:${ldap1ContainerWithToxiproxy.innerContainerMappedPort}").get),
           poolSize = 1,
           connectionTimeout = Refined.unsafeApply(5 seconds),
           requestTimeout = Refined.unsafeApply(5 seconds),
@@ -155,13 +150,34 @@ class LdapCircuitBreakerTests
       .getOrElse(throw new IllegalStateException("LDAP connection problem"))
   }
 
-  private def createCircuitBreakerDecoratedSimpleAuthenticationService() = {
-    new CircuitBreakerLdapAuthenticationServiceDecorator(
-      createSimpleAuthenticationService(),
-      CircuitBreakerConfig(
-        maxFailures = Refined.unsafeApply(2),
-        resetDuration = Refined.unsafeApply(0.5 second))
-    )
+  private def createHaAuthenticationService() = {
+    implicit val clock: Clock = Clock.systemUTC()
+    UnboundidLdapAuthenticationService
+      .create(
+        Name("my_ldap"),
+        ldapConnectionPoolProvider,
+        LdapConnectionConfig(
+          ConnectionMethod.SeveralServers(
+            NonEmptyList.of(
+              LdapHost.from(s"ldap://${SingletonLdapContainers.ldap1.ldapHost}:${SingletonLdapContainers.ldap1.ldapPort}").get,
+              LdapHost.from(s"ldap://${ldapContainerToStop.ldapHost}:${ldapContainerToStop.ldapPort}").get,
+            ),
+            HaMethod.RoundRobin
+          ),
+          poolSize = 1,
+          connectionTimeout = Refined.unsafeApply(5 seconds),
+          requestTimeout = Refined.unsafeApply(5 seconds),
+          trustAllCerts = false,
+          BindRequestUser.CustomUser(
+            Dn("cn=admin,dc=example,dc=com"),
+            PlainTextSecret("password")
+          ),
+          ignoreLdapConnectivityProblems = false
+        ),
+        UserSearchFilterConfig(Dn("ou=People,dc=example,dc=com"), CustomAttribute("uid"))
+      )
+      .runSyncUnsafe()
+      .getOrElse(throw new IllegalStateException("LDAP connection problem"))
   }
 
 }
