+++ core/src/test/scala/tech/beshu/ror/unit/acl/blocks/rules/http/UriRegexRuleTests.scala
@@ -21,7 +21,6 @@ import cats.implicits._
 import eu.timepit.refined.auto._
 import eu.timepit.refined.types.string.NonEmptyString
 import monix.execution.Scheduler.Implicits.global
-import org.scalamock.scalatest.MockFactory
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpec
 import tech.beshu.ror.accesscontrol.blocks.BlockContext.CurrentUserMetadataRequestBlockContext
@@ -40,7 +39,7 @@ import tech.beshu.ror.mocks.MockRequestContext
 import java.util.regex.Pattern
 import scala.util.Try
 
-class UriRegexRuleTests extends AnyWordSpec with MockFactory {
+class UriRegexRuleTests extends AnyWordSpec {
 
   "An UriRegexRule" should {
     "match" when {
